Fixed user authenticated

+++ TravelNotes/favorites/views.py
@@ -7,18 +7,23 @@ from notes.models import Note
 
 
 def show_favorites(request):
-	favorites = Favorite.objects.filter(user = request.user)
-	return render(request, 'favorites.html', {'favorites': favorites})
+	if request.user.is_authenticated:
+		favorites = Favorite.objects.filter(user = request.user)
+		return render(request, 'favorites.html', {'favorites': favorites})
+	else:
+		return redirect('../accounts/login')
 
 def add_favorite(request, fk):
-	post = get_object_or_404(Note, pk=fk)
-	user = request.user
-	favorite = Favorite(user = request.user, note = post)
-	favorite.save()
+	if request.user.is_authenticated:
+		post = get_object_or_404(Note, pk=fk)
+		favorite = Favorite(user = request.user, note = post)
+		favorite.save()
 	return redirect('post_detail', pk=fk)
 
+
 def remove_favorite(request, fk):
-	post = get_object_or_404(Note, pk=fk)
-	favorite = Favorite.objects.filter(user = request.user, note = post)
-	favorite.delete()
+	if request.user.is_authenticated:
+		post = get_object_or_404(Note, pk=fk)
+		favorite = Favorite.objects.filter(user = request.user, note = post)
+		favorite.delete()
 	return redirect('post_detail', pk=fk)
